fix: combine prev and next buttons with blogs existing styles for these kinds of links
